+++ controllers/monthly_sales_controller.py
@@ -1,15 +1,18 @@
-# controllers/komtrax_controller.py
-from fastapi import APIRouter, Depends, HTTPException, Path
+from io import BytesIO
+import json
+from fastapi import APIRouter, Depends, File, HTTPException, Path, UploadFile
+from fastapi.security import OAuth2PasswordBearer
 from sqlalchemy.orm import Session
 from typing import Annotated
 from starlette import status
 
+from helper.jwt_helper import JWTAuthHelper
 from models.dto.monthly_sales_request import MonthlySalesRequest
 from services.monthly_sales_services import MonthlySalesService
 from database import SessionLocal
 
 router = APIRouter()
-
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
 def get_db_monthly_sales():
     db = SessionLocal()
     try:
@@ -17,36 +20,80 @@ def get_db_monthly_sales():
     finally:
         db.close()
 
-db_dependency = Annotated[Session, Depends(get_db_monthly_sales)]
+# db_dependency = Annotated[Session, Depends(get_db_monthly_sales)]
 
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all_monthly_sales(db: db_dependency):
-    service = MonthlySalesService(db)
-    return service.read_all_monthly_sales_service(db)
+async def read_all_monthly_sales(token: str = Depends(oauth2_scheme), db: Session = Depends(get_db_monthly_sales)):
+    # Verify JWT token
+    current_user = JWTAuthHelper.get_current_user(token)
+    if current_user is not None:
+        # Initialize service and fetch data
+        service = MonthlySalesService(db)
+        return service.read_all_monthly_sales_service()
+    raise HTTPException(status_code=401, detail="Invalid Authentication Credentials")
 
 @router.get("/{monthly_sales_id}", status_code=status.HTTP_200_OK)
-async def read_komtrax_by_id(db: db_dependency, monthly_sales_id: int = Path(gt=0)):
-    service = MonthlySalesService(db)
-    monthly_sales_service_response = service.read_all_monthly_sales_by_id_service(db, monthly_sales_id)
-    if monthly_sales_service_response is not None:
-        return monthly_sales_service_response
-    raise HTTPException(status_code=404, detail='Data not found')
+async def read_monthly_sales_by_id(monthly_sales_id: int = Path(gt=0), token: str = Depends(oauth2_scheme), db: Session = Depends(get_db_monthly_sales)):
+    # Verify JWT token
+    current_user = JWTAuthHelper.get_current_user(token)
+    if current_user is not None:
+        # Initialize service and fetch data
+        service = MonthlySalesService(db)
+        monthly_sales_service_response = service.read_all_monthly_sales_by_id_service(monthly_sales_id)
+        if monthly_sales_service_response is not None:
+            return monthly_sales_service_response
+        raise HTTPException(status_code=404, detail='Data not found')
+    raise HTTPException(status_code=401, detail="Invalid Authentication Credentials")
 
 @router.post("/", status_code=status.HTTP_201_CREATED)
-async def create_monthly_sales(db: db_dependency, monthly_sales_request: MonthlySalesRequest):
+async def create_monthly_sales(monthly_sales_request: MonthlySalesRequest, token: str = Depends(oauth2_scheme), db: Session = Depends(get_db_monthly_sales)):
+    current_user = JWTAuthHelper.get_current_user(token)
+    if current_user is None:
+        raise HTTPException(status_code=401, detail="Invalid Authentication Credentials")
+
+    # Inisialisasi service dan buat data
     service = MonthlySalesService(db)
-    return service.create_monthly_sales_service(db, monthly_sales_request)
+    result = service.create_monthly_sales_service(monthly_sales_request)
+    return result
 
 @router.put("/{monthly_sales_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_monthly_sales(db: db_dependency, monthly_sales_id: int, monthly_sales_request: MonthlySalesRequest):
+async def update_monthly_sales(monthly_sales_id: int, monthly_sales_request: MonthlySalesRequest, token: str = Depends(oauth2_scheme), db: Session = Depends(get_db_monthly_sales)):
+    current_user = JWTAuthHelper.get_current_user(token)
+    if current_user is None:
+        raise HTTPException(status_code=401, detail="Invalid Authentication Credentials")
+
+    # Inisialisasi service dan perbarui data
     service = MonthlySalesService(db)
-    monthly_sales_service_response = service.update_monthly_sales_service(db, monthly_sales_id, monthly_sales_request)
-    if monthly_sales_service_response is None:
+    result = service.update_monthly_sales_service(monthly_sales_id, monthly_sales_request)
+    
+    if result is None:
         raise HTTPException(status_code=404, detail='Data not found')
     
 @router.delete("/{monthly_sales_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_monthly_sales(db:db_dependency, monthly_sales_id: int = Path(gt=0)):
+async def delete_monthly_sales(monthly_sales_id: int = Path(gt=0), token: str = Depends(oauth2_scheme), db: Session = Depends(get_db_monthly_sales)):
+    current_user = JWTAuthHelper.get_current_user(token)
+    if current_user is None:
+        raise HTTPException(status_code=401, detail="Invalid Authentication Credentials")
     service = MonthlySalesService(db)
     monthly_sales_service_response = service.delete_monthly_sales_service(db, monthly_sales_id)
     if monthly_sales_service_response is None:
         raise HTTPException(status_code=404, detail='Data not found')
+
+@router.post("/upload-excel/", status_code=status.HTTP_201_CREATED)
+async def upload_excel_file(file: UploadFile = File(...), token: str = Depends(oauth2_scheme), db: Session = Depends(get_db_monthly_sales)):
+    try:
+        current_user = JWTAuthHelper.get_current_user(token)
+        if current_user is None:
+            raise HTTPException(status_code=401, detail="Invalid Authentication Credentials")
+        config_file_path = "configfile.json"
+        with open(config_file_path, "r") as config_file:
+            config = json.load(config_file)
+        # service_name = "monthlySalesFile"
+        service = MonthlySalesService(db)
+        file_content = await file.read()
+        file_stream = BytesIO(file_content)
+        service.parse_and_create_monthly_sales(file_stream)
+        
+        return {"message": "File successfully processed and data saved to database"}
+    except Exception as e:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=str(e))
